requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.4.0
cffi==1.13.2
configparser==4.0.2
cryptography==2.8
defusedxml==0.6.0
Django==2.2.7
djangorestframework==3.11.0
grab==0.6.41
importlib-metadata==1.4.0
kombu==4.6.7
lxml==4.4.2
more-itertools==8.1.0
mysqlclient==1.4.6
pycparser==2.19
pycurl==7.43.0.1
pytils==0.3
pytz==2019.3
redis==3.3.11
selection==0.0.14
six==1.13.0
sqlparse==0.3.0
user-agent==0.1.9
vine==1.3.0
weblib==0.1.30
zipp==0.6.0